Extract EVChargerPool's StateTracker feature into a separate module
