+++ src/frequenz/sdk/timeseries/ev_charger_pool/__init__.py
@@ -3,7 +3,8 @@
 
 """Interactions with EV Chargers."""
 
-from ._ev_charger_pool import EVChargerPool, EVChargerPoolStates, EVChargerState
+from ._ev_charger_pool import EVChargerPool
+from ._state_tracker import EVChargerPoolStates, EVChargerState
 
 __all__ = [
     "EVChargerPool",
